backend/requirements.txt
asgiref
Django
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
# mysql-connector-python==8.3.0
mysqlclient
gunicorn
openpyxl==3.1.2
pillow
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
whitenoise==6.6.0
python-decouple